+++ src/pages/CartPage/CartPage.css
@@ -1,0 +1,12 @@
+.main-container {
+  width: 100vw;
+  height: 100vh;
+  display: grid;
+  place-items: center;
+}
+
+.sub-container {
+  margin: 0 auto;
+  max-width: 1280px;
+  width: 90%;
+}
